pyproject.toml
[tool.poetry]
name = "skylet"
version = "0.1.0"
description = "A bot version of SkyLissh to do hard taks"
authors = ["Alisson Lopez <alisson.hdz.lopez@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
"discord.py" = "^1.7.3"
tabulate = "^0.8.9"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
flake8 = "^4.0.1"
black = "^22.1.0"
isort = "^5.10.1"
autoflake = "^1.4"
"discord.py-stubs" = "^1.7.3"
types-tabulate = "^0.8.6"

[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | __pycache__
  | _build
  | buck_out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 90
include_trailing_comma = true
force_grid_wrap = 0

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
